+++ cmd/get/nodepools/printer.go
@@ -57,6 +57,4 @@ func (r *runner) printOutput(npResource nodepool.Resource) error {
 
 func (r *runner) printNoResourcesOutput() {
 	_, _ = fmt.Fprintf(r.stdout, "No node pools found.\n")
-	_, _ = fmt.Fprintf(r.stdout, "To create a node pool, please check\n\n")
-	_, _ = fmt.Fprintf(r.stdout, "  kubectl gs template nodepool --help\n")
 }