+++ cmd/update/cluster/runner_test.go
@@ -6,15 +6,14 @@ import (
 	"fmt"
 	"testing"
 
-	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v6/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/k8sclient/v8/pkg/k8sclienttest"
 	"github.com/giantswarm/microerror"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
-	capi "sigs.k8s.io/cluster-api/api/core/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake" //nolint:staticcheck
 
+	"github.com/giantswarm/kubectl-gs/v5/internal/key"
 	"github.com/giantswarm/kubectl-gs/v5/internal/label"
 	"github.com/giantswarm/kubectl-gs/v5/pkg/commonconfig"
 	"github.com/giantswarm/kubectl-gs/v5/pkg/data/domain/cluster"
@@ -69,47 +68,43 @@ func Test_run(t *testing.T) {
 	}
 }
 
-func newCluster(name, namespace, targetRelease string) *capi.Cluster {
-	c := &capi.Cluster{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "cluster.x-k8s.io/v1beta1",
-			Kind:       "Cluster",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-			Labels: map[string]string{
-				capi.ClusterNameLabel: name,
-				label.ReleaseVersion:  "16.0.1",
-			},
-			Annotations: map[string]string{
-				"cluster.giantswarm.io/description": "fake-cluster",
+func newCluster(name, namespace, targetRelease string) *unstructured.Unstructured {
+	return &unstructured.Unstructured{
+		Object: map[string]interface{}{
+			"apiVersion": "cluster.x-k8s.io/v1beta1",
+			"kind":       "Cluster",
+			"metadata": map[string]interface{}{
+				"name":      name,
+				"namespace": namespace,
+				"labels": map[string]interface{}{
+					key.ClusterNameLabel: name,
+					label.ReleaseVersion: "16.0.1",
+				},
+				"annotations": map[string]interface{}{
+					"cluster.giantswarm.io/description": "fake-cluster",
+				},
 			},
 		},
 	}
-
-	return c
 }
 
-func newAWSCluster(name, namespace, targetRelease string) *infrastructurev1alpha3.AWSCluster {
-	c := &infrastructurev1alpha3.AWSCluster{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "infrastructure.giantswarm.io/v1alpha3",
-			Kind:       "AWSCluster",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-			Labels: map[string]string{
-				label.Cluster: name,
-			},
-			Annotations: map[string]string{
-				"cluster.giantswarm.io/description": "fake-cluster",
+func newAWSCluster(name, namespace, targetRelease string) *unstructured.Unstructured {
+	return &unstructured.Unstructured{
+		Object: map[string]interface{}{
+			"apiVersion": "infrastructure.giantswarm.io/v1alpha3",
+			"kind":       "AWSCluster",
+			"metadata": map[string]interface{}{
+				"name":      name,
+				"namespace": namespace,
+				"labels": map[string]interface{}{
+					label.Cluster: name,
+				},
+				"annotations": map[string]interface{}{
+					"cluster.giantswarm.io/description": "fake-cluster",
+				},
 			},
 		},
 	}
-
-	return c
 }
 
 func newClusterService(t *testing.T, object ...runtime.Object) *cluster.Service {